Change API route files paths to be more explicit
Also two infinitesimal changes to the dash site
