+++ app/client/pages/api/graphite/logs/[jobId].js
@@ -2,8 +2,7 @@ import { GRAPHITE_LOGS_URL } from '../../../../utils/const';
 import handleRequestResponse from '../../../../utils/handleRequestResponse';
 
 export default async (req, res) => {
-  const { slug } = req.query;
-  const jobId = slug[0];
+  const { jobId } = req.query;
   const url = GRAPHITE_LOGS_URL(jobId);
   const method = 'GET';
   const checkAuth = false;
